refactor(sidebar): common sidebar for all pages
- add latex names of variables e, i and N
