+++ pages/3_Savings_estimation.py
@@ -1,66 +1,14 @@
-import json
-
 import numpy as np
 import pandas as pd
 import streamlit as st
 
-from src.utils import get_dataset_parameters, calculate_metrics, defined_metrics
-
-with open("data/datasets.json", "r") as f:
-    datasets = json.load(f)
-
-# Sidebar
-st.sidebar.write("### Dataset parameters")
-
-emp = st.sidebar.empty()
-dataset_type = emp.selectbox(
-    label="Select a dataset type", options=datasets.keys(), key="dataset_picker"
+from src.utils import (
+    calculate_metrics,
+    defined_metrics,
+    draw_sidebar,
 )
 
-if dataset_type == "Custom":
-    _dataset_size = st.session_state.dataset_size
-    _i_percentage = st.session_state.i_percentage
-    _i = int(_dataset_size * _i_percentage / 100)
-    _e = _dataset_size - _i
-else:
-    _dataset_size, _i, _e, _i_percentage = get_dataset_parameters(
-        dataset_type=dataset_type
-    )
-
-
-def check_dataset_size():
-    if _dataset_size != st.session_state.dataset_size:
-        st.session_state["dataset_picker"] = "Custom"
-
-
-def check_i_percentage():
-    if _i_percentage != st.session_state.i_percentage:
-        st.session_state["dataset_picker"] = "Custom"
-
-
-dataset_size = st.sidebar.slider(
-    "Dataset size",
-    100,
-    5000,
-    _dataset_size,
-    50,
-    key="dataset_size",
-    on_change=check_dataset_size,
-)
-i_percentage = st.sidebar.slider(
-    "Percentage of relevant documents (includes)",
-    1.0,
-    99.0,
-    _i_percentage,
-    1.0,
-    key="i_percentage",
-    on_change=check_i_percentage,
-)
-
-i = int(dataset_size * i_percentage / 100)
-e = dataset_size - i
-st.sidebar.write("Number of relevant documents (includes): ", i)
-st.sidebar.write("Number of non-relevant documents (excludes): ", e)
+e, i, dataset_size = draw_sidebar()
 
 st.title("Estimation of time and money savings depending on evaluation measures values")
